requirements.txt
fastapi
uvicorn[standard]
pyodbc
pydantic
python-dotenv
httpx
pytest
python-multipart